+++ src/NotFoundPage.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function NotFoundPage() {
+  return (
+    <h2>404 Not Found</h2>
+  );
+}
